+++ navigators/stackAccNav.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {createStackNavigator} from '@react-navigation/stack'
 import {AccountScreen} from '../screens/Account';
-import {SettingsScreen} from '../screens/AccountScreens/Settings'; 
+import {SettingsScreen} from '../screens/AccountScreens/Settings/Settings'; 
 import {ContactScreen} from '../screens/AccountScreens/Contact'; 
 import {AboutScreen} from '../screens/AccountScreens/About';
 import {LogInScreen} from '../screens/LogIn';
