--- screens/AccountScreens/Settings.js
@@ -1,22 +1,0 @@
-import React from 'react'; 
-import { Text, View, StyleSheet, Button} from 'react-native'; 
-
-const SettingsScreen = props => { 
-  console.log(props)
-    return (
-        <View style={styles.screen}>
-            <Text>Settings Screen</Text>
-            <Button title="Go Back" onPress={() => props.navigation.goBack()}/>
-        </View>   
-    )
-}
-
-const styles = StyleSheet.create({
-  screen: { 
-    flex: 1, 
-    justifyContent: 'center',
-    alignItems: 'center'
-  }
-})
-
-export {SettingsScreen}
